Extract trigger action logic to shared file
